1.6.1.5
ODIN has been updated to support reconfiguration right from the
application itself
without the need for a secondary batch file. Simply press the key combo
CTRL+SHIFT+R.
Added the binary to handle ODIN reconfiguration as a payload of OSIRiS
so it can live
in the C:\profiles\ directory with all other on-disk OSIRiS content.